moved mapping and validation to a separate project
