+++ Apsitvarkom.ModelActions/Validation/ObjectIdentifyRequestValidator.cs
@@ -1,0 +1,12 @@
+﻿using Apsitvarkom.Models.Public;
+using FluentValidation;
+
+namespace Apsitvarkom.ModelActions.Validation;
+
+public class ObjectIdentifyRequestValidator : AbstractValidator<ObjectIdentifyRequest>
+{
+    public ObjectIdentifyRequestValidator()
+    {
+        RuleFor(l => l.Id).NotNull();
+    }
+}
