+++ Apsitvarkom.Models/Apsitvarkom.Models.csproj
@@ -6,9 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <PackageReference Include="AutoMapper" Version="12.0.0" />
-    <PackageReference Include="FluentValidation" Version="11.2.2" />
-  </ItemGroup>
-
 </Project>
