+++ Apsitvarkom.Models/Public/CleaningEventCreateRequest.cs
@@ -1,6 +1,4 @@
-﻿using FluentValidation;
-
-namespace Apsitvarkom.Models.Public;
+﻿namespace Apsitvarkom.Models.Public;
 
 public class CleaningEventCreateRequest
 {
@@ -9,13 +7,4 @@ public class CleaningEventCreateRequest
     public DateTime? StartTime { get; set; }
 
     public string? Notes { get; set; }
-}
-
-public class CleaningEventCreateRequestValidator : AbstractValidator<CleaningEventCreateRequest>
-{
-    public CleaningEventCreateRequestValidator()
-    {
-        RuleFor(l => l.StartTime).NotNull().GreaterThan(DateTime.UtcNow);
-        RuleFor(l => l.PollutedLocationId).NotNull();
-    }
 }
