+++ Apsitvarkom.Models/Public/CleaningEventUpdateRequest.cs
@@ -1,6 +1,4 @@
-﻿using FluentValidation;
-
-namespace Apsitvarkom.Models.Public;
+﻿namespace Apsitvarkom.Models.Public;
 
 public class CleaningEventUpdateRequest
 {
@@ -9,13 +7,4 @@ public class CleaningEventUpdateRequest
     public DateTime? StartTime { get; set; }
 
     public string? Notes { get; set; }
-}
-
-public class CleaningEventUpdateRequestValidator : AbstractValidator<CleaningEventUpdateRequest>
-{
-    public CleaningEventUpdateRequestValidator()
-    {
-        RuleFor(l => l.Id).NotNull();
-        RuleFor(l => l.StartTime).NotNull().GreaterThan(DateTime.UtcNow);
-    }
 }
