+++ Apsitvarkom.Models/Public/ObjectIdentifyRequest.cs
@@ -1,16 +1,6 @@
-﻿using FluentValidation;
-
-namespace Apsitvarkom.Models.Public;
+﻿namespace Apsitvarkom.Models.Public;
 
 public class ObjectIdentifyRequest
 {
     public Guid? Id { get; set; }
-}
-
-public class ObjectIdentifyRequestValidator : AbstractValidator<ObjectIdentifyRequest>
-{
-    public ObjectIdentifyRequestValidator()
-    {
-        RuleFor(l => l.Id).NotNull();
-    }
 }
