+++ Apsitvarkom.Models/Public/PollutedLocationCreateRequest.cs
@@ -1,5 +1,4 @@
-﻿using FluentValidation;
-using static Apsitvarkom.Models.PollutedLocation;
+﻿using static Apsitvarkom.Models.PollutedLocation;
 
 namespace Apsitvarkom.Models.Public;
 
@@ -12,16 +11,4 @@ public class PollutedLocationCreateRequest
     public SeverityLevel? Severity { get; set; }
 
     public string? Notes { get; set; }
-}
-
-public class PollutedLocationCreateRequestValidator : AbstractValidator<PollutedLocationCreateRequest>
-{
-    public PollutedLocationCreateRequestValidator(IValidator<LocationCreateRequest> locationCreateRequestValidator)
-    {
-        RuleFor(l => l.Radius).NotNull().GreaterThanOrEqualTo(1);
-        RuleFor(l => l.Severity).NotNull();
-
-        RuleFor(l => l.Location).NotNull();
-        RuleFor(l => l.Location!).SetValidator(locationCreateRequestValidator).When(l => l.Location is not null);
-    }
 }
