+++ Apsitvarkom.Models/Public/PollutedLocationResponse.cs
@@ -1,5 +1,4 @@
-﻿using FluentValidation;
-using static Apsitvarkom.Models.PollutedLocation;
+﻿using static Apsitvarkom.Models.PollutedLocation;
 
 namespace Apsitvarkom.Models.Public;
 
@@ -20,15 +19,4 @@ public class PollutedLocationResponse
     public string? Notes { get; set; }
 
     public List<CleaningEventResponse> Events { get; set; } = new();
-}
-
-public class PollutedLocationResponseValidator : AbstractValidator<PollutedLocationResponse>
-{
-    public PollutedLocationResponseValidator(IValidator<LocationResponse> locationResponseValidator)
-    {
-        RuleFor(l => l.Location).SetValidator(locationResponseValidator);
-
-        RuleFor(l => l.Radius).GreaterThanOrEqualTo(1);
-        RuleFor(l => l.Progress).InclusiveBetween(0, 100);
-    }
 }
