+++ Apsitvarkom.Models/Public/PollutedLocationUpdateRequest.cs
@@ -1,5 +1,4 @@
-﻿using FluentValidation;
-using static Apsitvarkom.Models.PollutedLocation;
+﻿using static Apsitvarkom.Models.PollutedLocation;
 
 namespace Apsitvarkom.Models.Public;
 
@@ -12,14 +11,4 @@ public class PollutedLocationUpdateRequest
     public SeverityLevel? Severity { get; set; }
 
     public string? Notes { get; set; }
-}
-
-public class PollutedLocationUpdateRequestValidator : AbstractValidator<PollutedLocationUpdateRequest>
-{
-    public PollutedLocationUpdateRequestValidator()
-    {
-       RuleFor(l => l.Id).NotNull();
-       RuleFor(l => l.Radius).NotNull().GreaterThanOrEqualTo(1);
-       RuleFor(l => l.Severity).NotNull();
-    }
 }
