+++ Apsitvarkom.UnitTests/Api/Controllers/CleaningEventControllerTests.cs
@@ -1,8 +1,9 @@
 ﻿using System.Linq.Expressions;
 using Apsitvarkom.Api.Controllers;
 using Apsitvarkom.DataAccess;
+using Apsitvarkom.ModelActions.Mapping;
+using Apsitvarkom.ModelActions.Validation;
 using Apsitvarkom.Models;
-using Apsitvarkom.Models.Mapping;
 using Apsitvarkom.Models.Public;
 using AutoMapper;
 using Microsoft.AspNetCore.Http;
