+++ Apsitvarkom.UnitTests/ModelActions/Validation/CleaningEventCreateRequestValidationTests.cs
@@ -1,7 +1,8 @@
-﻿using Apsitvarkom.Models.Public;
+﻿using Apsitvarkom.ModelActions.Validation;
+using Apsitvarkom.Models.Public;
 using FluentValidation;
 
-namespace Apsitvarkom.UnitTests.Models.Validation;
+namespace Apsitvarkom.UnitTests.ModelActions.Validation;
 
 public class CleaningEventCreateRequestValidationTests
 {
@@ -64,7 +65,7 @@ public class CleaningEventCreateRequestValidationTests
 
         Assert.That(result, Is.Not.Null);
         // Only 1 thing must error in the invalid input
-        Assert.That(result.Errors.Count, Is.EqualTo(1));
+        Assert.That(result.Errors, Has.Count.EqualTo(1));
         Assert.That(result.IsValid, Is.False);
     }
 }
