+++ Apsitvarkom.UnitTests/ModelActions/Validation/PollutedLocationUpdateRequestValidationTests.cs
@@ -1,8 +1,9 @@
-﻿using Apsitvarkom.Models.Public;
+﻿using Apsitvarkom.ModelActions.Validation;
 using Apsitvarkom.Models;
+using Apsitvarkom.Models.Public;
 using FluentValidation;
 
-namespace Apsitvarkom.UnitTests.Models.Validation;
+namespace Apsitvarkom.UnitTests.ModelActions.Validation;
 
 public class PollutedLocationUpdateRequestValidatorTests
 {
@@ -72,7 +73,7 @@ public class PollutedLocationUpdateRequestValidatorTests
         var result = await Validator.ValidateAsync(input);
 
         Assert.That(result, Is.Not.Null);
-        Assert.That(result.Errors.Count, Is.EqualTo(1));
+        Assert.That(result.Errors, Has.Count.EqualTo(1));
         Assert.That(result.IsValid, Is.False);
     }
 }
