+++ Apsitvarkom.sln
@@ -15,6 +15,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apsitvarkom.Configuration",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apsitvarkom.IntegrationTests", "Apsitvarkom.IntegrationTests\Apsitvarkom.IntegrationTests.csproj", "{3FD00F6F-CE03-471C-B057-85B501A5D743}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apsitvarkom.ModelActions", "Apsitvarkom.ModelActions\Apsitvarkom.ModelActions.csproj", "{51998D7B-E5D3-42FF-97A0-456C40E7AF57}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -45,6 +47,10 @@ Global
 		{3FD00F6F-CE03-471C-B057-85B501A5D743}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3FD00F6F-CE03-471C-B057-85B501A5D743}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3FD00F6F-CE03-471C-B057-85B501A5D743}.Release|Any CPU.Build.0 = Release|Any CPU
+		{51998D7B-E5D3-42FF-97A0-456C40E7AF57}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{51998D7B-E5D3-42FF-97A0-456C40E7AF57}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{51998D7B-E5D3-42FF-97A0-456C40E7AF57}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{51998D7B-E5D3-42FF-97A0-456C40E7AF57}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
